backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
pyannote.audio==3.1.1
anthropic==0.18.0
pydub==0.25.1
ffmpeg-python==0.2.0
httpx==0.26.0
python-dotenv==1.0.0
torch>=2.0.0
torchaudio>=2.0.0
aiofiles==23.2.1
